requirements.txt
langchain==0.1.15
python-dotenv==1.0.0
python-telegram-bot==20.7
redis==5.0.1
gigachat
pytesseract~=0.3.10
pillow~=10.3.0
requests~=2.31.0
psycopg2~=2.9.9
peewee~=3.17.3